backend/pyproject.toml
[project]
name = "processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.1",
    "deep-translator>=1.11.4",
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "huggingface-hub>=0.36.0",
    "markdown>=3.10",
    "openai>=2.15.0",
    "openai-whisper>=20250625",
    "pdfkit>=1.0.0",
    "pydantic>=2.12.5",
    "pydub>=0.25.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "uvicorn>=0.40.0",
    "xhtml2pdf>=0.2.17",
]
